\name{pom-package}
\alias{pom-package}
\alias{pom}
\docType{package}
\title{POM - Patch Occupancy Models}
\description{The package implements models to analyze site occupancy data with detection error.}
\details{
\tabular{ll}{
Package: \tab pom\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-17\cr
License: \tab  GNU General Public License\cr
LazyData: \tab yes\cr
}
List of routines:

\tabular{ll}{
\code{\link{siteocc}} \tab Routine for Estimating the Patch Occupancy Model. \cr
\code{\link{print}} \tab Printing Patch Occupancy Model Fits. \cr
\code{\link{F.2nd.deriv}} \tab Generic Routine for calculating the hessian matrix of a function. \cr
}
}
\author{
Fawn Hornsby, Ryan Nielson, and Trent McDonald \url{www.west-inc.com}
	
Maintainer: Fawn Hornsby \email{fhornsby@west-inc.com}
}
\references{
Casella, G., and Berger, R. L. (2002). Statistical inference, Duxbury Press.

MacKenzie, D.I., Nichols, J.D., Lachman, G. B., Droege, S., Royle, J. A., and Langtimm, C. A., 2002. Estimating site occupancy rates when detection probabilities are less than one. Ecology, 83:2248-2255.

MacKenzie, D.I., Nichols, J.D., Royle, J.A., and Pollock, K.H. (2006), Occupancy Estimation and Modeling: Inferring Patterns and Dynamics of Species Occurrence, Academic Press, Burlington, MA.

Royle, J.A., 2006. Site occupancy models with heterogeneous detection probabilities. Biometrics 62:97-102.
}
\keyword{ package }